backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
python-dotenv
alembic
aiosqlite
# Testing
pytest
pytest-asyncio
httpx